solution monte carlo

+++ monte_carlo.c
@@ -13,7 +13,7 @@ for(int i=1; i<=n; i++){
     if (((x*x)+ (y*y))<=1){circle++;}
     else {square++;}
 } square = square + circle;
- float pi = 4*circle/square;
+ float pi = 4* (float)circle/square;
 return pi;
 }
